Archi modifications easier to read and opti for the moment + Renaming functions
