+++ app/controller/index.js
@@ -2,9 +2,11 @@
 const userController = require("./user");
 const crewController = require("./crew");
 const eventController = require("./event");
+const friendController = require("./friend");
 
 module.exports = {
   userController,
   crewController,
   eventController,
+  friendController,
 };
