+++ app/controller/user.js
@@ -116,349 +116,6 @@ const userController = {
   async login(req, res, next) {
     //! TO DO
   },
-
-  /**
-   * ! GET ALL FRIENDS
-   * Method to get all friend
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async getAllFriends(req, res, next) {
-    const user = req.params.userId;
-    console.log(user);
-    const { error, result } = await userDatamapper.getAllFriends(user);
-    if (error) {
-      next(error);
-    } else {
-      res.json(result);
-    }
-  },
-
-  /**
-   * ! GET ONE FRIEND
-   * Method to get one friend
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async getOneFriend(req, res, next) {
-    console.log(req.params);
-    const userId = req.params.userId;
-    const friendId = req.params.friendId;
-    const { error, result } = await userDatamapper.getOneFriend(
-      userId,
-      friendId
-    );
-    if (error) {
-      next(error);
-    } else {
-      res.json(result);
-    }
-  },
-
-  /**
-   * ! ADD ONE FRIEND
-   * Method to add one friend
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async addOneFriend(req, res, next) {
-    const userId = req.params.userId;
-    const friendId = req.body.friendId;
-    const { error, result } = await userDatamapper.addOneFriend(
-      userId,
-      friendId
-    );
-    if (error) {
-      next(error);
-    } else {
-      res.json(result);
-    }
-  },
-
-  /**
-   * ! DELETE ONE FRIEND
-   * Method to delete one friend
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async deleteOneFriend(req, res, next) {
-    const userId = req.params.userId;
-    const friendId = req.body.friendId;
-    const { error, result } = await userDatamapper.deleteOneFriend(
-      userId,
-      friendId
-    );
-    if (error) {
-      next(error);
-    } else {
-      res.json(result);
-    }
-  },
-
-  // ! CREWS
-
-  /**
-   * ! GET ALL CREWS
-   * Method to get all crews
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async getAllCrews(req, res, next) {
-    const user = req.params.userId;
-    console.log(user);
-    const { error, result } = await userDatamapper.getAllCrews(user);
-    if (error) {
-      next(error);
-    } else {
-      res.json(result);
-    }
-  },
-
-  /**
-   * ! GET ONE CREW
-   * Method to get one crew
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async getOneCrew(req, res, next) {
-    console.log(req.params);
-    const userId = req.params.userId;
-    const crewId = req.params.crewId;
-    const { error, result } = await userDatamapper.getOneCrew(userId, crewId);
-    if (error) {
-      next(error);
-    } else {
-      res.json(result);
-    }
-  },
-
-  /**
-   * ! ADD ONE CREW
-   * Method to add one crew
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async addOneCrew(req, res, next) {
-    const userId = req.params.userId;
-    const crewName = req.body.crew_name;
-    const crewPicture = req.body.crew_picture;
-    const crewMembers = req.body.added_friends;
-
-    console.log(
-      "controllers params :",
-      userId,
-      crewName,
-      crewPicture,
-      crewMembers
-    );
-
-    const { error, result } = await userDatamapper.addOneCrew(
-      userId,
-      crewName,
-      crewPicture,
-      crewMembers
-    );
-    if (error) {
-      next(error);
-    } else {
-      res.json(result);
-    }
-  },
-
-  /**
-   * ! MODIFY ONE
-   * Method to modify a crew
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async modifyOneCrew(req, res, next) {
-    const userId = req.params.userId;
-    const crewInfo = req.body;
-    console.log("controller : ", userId, crewInfo);
-    if (req.body.userId == req.params.userId) {
-      const { error, result } = await userDatamapper.modifyOneCrew(crewInfo);
-      if (error) {
-        next(error);
-      } else {
-        res.json(result);
-      }
-    } else {
-      const err = new APIError("Acces denied", 404);
-      next(err);
-    }
-  },
-
-  /**
-   * ! DELETE ONE CREW
-   * Method to delete one crew
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async deleteOneCrew(req, res, next) {
-    const userId = parseInt(req.params.userId);
-    const userOwner = req.body.userId;
-    const crewId = req.body.crew_id_param;
-    console.log(userId, userOwner, crewId);
-    if (userId !== userOwner) {
-      return res
-        .status(403)
-        .json({ error: "Unauthorized access. userId must match userOwner." });
-    }
-    const { error, result } = await userDatamapper.deleteOneCrew(
-      userId,
-      crewId
-    );
-    if (error) {
-      next(error);
-    } else {
-      res.json(result);
-    }
-  },
-
-  // ! EVENTS
-
-  /**
-   * ! GET ALL EVENTS
-   * Method to get all events
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async getAllEvents(req, res, next) {
-    const user = req.params.userId;
-    console.log(user);
-    const { error, result } = await userDatamapper.getAllEvents(user);
-    if (error) {
-      next(error);
-    } else {
-      res.json(result);
-    }
-  },
-
-  /**
-   * ! GET ONE EVENT
-   * Method to get one crew
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async getOneEvent(req, res, next) {
-    console.log(req.params);
-    const userId = req.params.userId;
-    const eventId = req.params.eventId;
-    const { error, result } = await userDatamapper.getOneEvent(userId, eventId);
-    if (error) {
-      next(error);
-    } else {
-      res.json(result);
-    }
-  },
-
-  /**
-   * ! ADD ONE EVENT
-   * Method to add one crew
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async addOneEvent(req, res, next) {
-    const userId = req.params.userId;
-    const {
-      theme,
-      date,
-      time,
-      place,
-      nb_people,
-      invited_users_ids,
-      invited_crews_ids,
-    } = req.body;
-
-    console.log(
-      "controllers params :",
-      userId,
-      theme,
-      date,
-      time,
-      place,
-      nb_people,
-      invited_users_ids,
-      invited_crews_ids
-    );
-
-    const { error, result } = await userDatamapper.addOneEvent(
-      userId,
-      theme,
-      date,
-      time,
-      place,
-      nb_people,
-      invited_users_ids,
-      invited_crews_ids
-    );
-    if (error) {
-      next(error);
-    } else {
-      res.json(result);
-    }
-  },
-
-  /**
-   * ! MODIFY ONE
-   * Method to modify a event
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async modifyOneEvent(req, res, next) {
-    const userId = req.params.userId;
-    const eventInfo = req.body;
-    console.log("controller : ", userId, eventInfo);
-    if (req.body.userId == req.params.userId) {
-      const { error, result } = await userDatamapper.modifyOneEvent(eventInfo);
-      if (error) {
-        next(error);
-      } else {
-        res.json(result);
-      }
-    } else {
-      const err = new APIError("Acces denied", 404);
-      next(err);
-    }
-  },
-
-  /**
-   * ! DELETE ONE EVENT
-   * Method to delete one event
-   * @param {*} req
-   * @param {*} res
-   * @param {*} next
-   */
-  async deleteOneEvent(req, res, next) {
-    const userId = parseInt(req.params.userId);
-    const userOwner = req.body.userId;
-    const eventId = req.body.eventId;
-    console.log(userId, userOwner, eventId);
-    if (userId !== userOwner) {
-      return res
-        .status(403)
-        .json({ error: "Unauthorized access. userId must match userOwner." });
-    }
-    const { error, result } = await userDatamapper.deleteOneEvent(eventId);
-    if (error) {
-      next(error);
-    } else {
-      res.json(result);
-    }
-  },
 };
 
 module.exports = userController;
