+++ app/model/index.js
@@ -2,9 +2,11 @@
 const userDatamapper = require("./user");
 const crewDatamapper = require("./crew");
 const eventDatamapper = require("./event");
+const friendDatamapper = require("./friend");
 
 module.exports = {
   userDatamapper,
   crewDatamapper,
   eventDatamapper,
+  friendDatamapper,
 };
