+++ script/sql/fonction/delete_fonction.sql
@@ -1,20 +1,24 @@
 -- DROP ALL EXISTING FUNCTION
-
+-- USER
 DROP FUNCTION IF EXISTS web.get_all_users;
 DROP FUNCTION IF EXISTS web.get_one_user;
 DROP FUNCTION IF EXISTS web.insert_user;
 DROP FUNCTION IF EXISTS web.update_user;
 DROP FUNCTION IF EXISTS web.delete_user;
-DROP FUNCTION IF EXISTS web.get_all_friends;
-DROP FUNCTION IF EXISTS web.get_one_friend;
-DROP FUNCTION IF EXISTS web.add_friend_to_user;
-DROP FUNCTION IF EXISTS web.get_user_crews_with_users;
+-- FRIEND
+DROP FUNCTION IF EXISTS web.get_user_all_friends;
+DROP FUNCTION IF EXISTS web.get_user_one_friend;
+DROP FUNCTION IF EXISTS web.insert_user_friend;
+DROP FUNCTION IF EXISTS web.delete_user_friend;
+-- CREW
+DROP FUNCTION IF EXISTS web.get_user_all_crews;
 DROP FUNCTION IF EXISTS web.get_user_one_crew;
-DROP FUNCTION IF EXISTS web.create_crew_for_users;
-DROP FUNCTION IF EXISTS web.delete_crew_and_links;
-DROP FUNCTION IF EXISTS web.get_user_events_with_invitations;
-DROP FUNCTION IF EXISTS web.create_event_for_users;
-DROP FUNCTION IF EXISTS web.edit_user_event;
-DROP FUNCTION IF EXISTS web.delete_event_by_id;
-DROP FUNCTION IF EXISTS web.update_event_details;
-DROP FUNCTION IF EXISTS web.update_crew;
+DROP FUNCTION IF EXISTS web.insert_user_crew;
+DROP FUNCTION IF EXISTS web.update_user_crew;
+DROP FUNCTION IF EXISTS web.delete_user_crew;
+-- EVENT
+DROP FUNCTION IF EXISTS web.get_user_all_events;
+DROP FUNCTION IF EXISTS web.get_user_one_event;
+DROP FUNCTION IF EXISTS web.insert_user_event;
+DROP FUNCTION IF EXISTS web.update_user_event;
+DROP FUNCTION IF EXISTS web.delete_user_event;
